refactor: schemas.py 파일명 도메인 역할 표시
