+++ app/domains/auth/controller/auth_router.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Cookie, Depends, HTTPException, Request, Response
 
 from app.domains.auth.dependency.auth_dependencies import get_auth_service, get_current_user
 from app.domains.auth.models import User
-from app.domains.auth.schema.schemas import (
+from app.domains.auth.schema.auth_schemas import (
     AuthTokenResponse,
     LoginRequest,
     MessageResponse,
