+++ app/domains/progress/controller/student_progress_router.py
@@ -6,7 +6,7 @@ from app.domains.auth.models import User
 from app.domains.content.dependency.catalog_dependencies import get_content_catalog_service
 from app.domains.content.service.catalog_service import ContentCatalogService
 from app.domains.progress.dependency.learning_record_dependencies import get_learning_record_service
-from app.domains.progress.schema.schemas import (
+from app.domains.progress.schema.learning_record_schemas import (
     LearningRecordResponse,
     LearningRecordsResponse,
     StudentProgressResponse,
