+++ tests/structure/test_domain_structure.py
@@ -12,7 +12,7 @@ def test_domain_packages_are_importable():
         "app.domains.auth.models",
         "app.domains.auth.repository.auth_repository",
         "app.domains.auth.controller.auth_router",
-        "app.domains.auth.schema.schemas",
+        "app.domains.auth.schema.auth_schemas",
         "app.domains.auth.service.auth_service",
         "app.domains.classroom.controller.teacher_class_router",
         "app.domains.classroom.controller.student_class_router",
@@ -21,8 +21,8 @@ def test_domain_packages_are_importable():
         "app.domains.classroom.models",
         "app.domains.classroom.repository.classroom_repository",
         "app.domains.classroom.service.classroom_service",
-        "app.domains.classroom.schema.schemas",
-        "app.domains.agent.schema.schemas",
+        "app.domains.classroom.schema.classroom_schemas",
+        "app.domains.agent.schema.agent_schemas",
         "app.domains.agent.service.agent_service",
         "app.domains.instruction.controller.student_router",
         "app.domains.instruction.controller.teacher_router",
@@ -30,7 +30,7 @@ def test_domain_packages_are_importable():
         "app.domains.instruction.validation.validation",
         "app.domains.instruction.models",
         "app.domains.instruction.repository.assignment_repository",
-        "app.domains.instruction.schema.schemas",
+        "app.domains.instruction.schema.instruction_schemas",
         "app.domains.instruction.service.instruction_service",
         "app.domains.content.controller.catalog_router",
         "app.domains.content.stage1.router",
@@ -44,17 +44,17 @@ def test_domain_packages_are_importable():
         "app.domains.content",
         "app.domains.content.models",
         "app.domains.content.repository.catalog_repository",
-        "app.domains.content.schema.schemas",
+        "app.domains.content.schema.catalog_schemas",
         "app.domains.content.service.catalog_service",
         "app.domains.progress.models",
         "app.domains.progress.repository.learning_record_repository",
-        "app.domains.progress.schema.schemas",
+        "app.domains.progress.schema.learning_record_schemas",
         "app.domains.progress.service.learning_record_service",
         "app.domains.content.stage3.schemas",
         "app.domains.content.stage3.service",
         "app.domains.agent",
         "app.domains.agent.controller.agent_router",
-        "app.domains.agent.schema.schemas",
+        "app.domains.agent.schema.agent_schemas",
         "app.infrastructure.rag",
         "app.infrastructure.rag.retriever",
         "app.infrastructure.rag.schemas",
